First pass at the booking endpoint.
BookingDao - Had to add a function to update an existing booking.

BookingRequest - Added this because even though the specs don't require additional information, in a real world scenario, there probably would be additional information needed, and with more time, we could expand out to keeping a guest list with information on each guest for each booking.

ReservationLedger - Added the function to determine the best room. Basically, from both a profit and an availability perspective, it's best to fill each room up as much as possible, because that leaves the other rooms open and flexible for other requests. Basically, put people in the rooms that they fit into best. Ergh, I'll figure out how to better word this later.

Application - New post endpoint~~~~